melhorias na classe carrinho e venda

+++ src/main/kotlin/com/grocery/jumarket/dto/NewProdutoDTO.kt
@@ -1,8 +1,10 @@
 package com.grocery.jumarket.dto
 
+import java.math.BigDecimal
+
 data class NewProdutoDTO(
         val nome: String,
         val unidadeDeMedida: String,
-        val precoUnitario: Double,
+        val precoUnitario: BigDecimal,
         val categoriaId: Long
 )
