+++ src/main/kotlin/com/grocery/jumarket/dto/ProdutoDTO.kt
@@ -1,9 +1,11 @@
 package com.grocery.jumarket.dto
 
+import java.math.BigDecimal
+
 data class ProdutoDTO(
         val id: Long?,
         val nome: String,
         val unidadeDeMedida: String,
-        val precoUnitario: Double,
+        val precoUnitario: BigDecimal,
         val categoriaId: Long?
 )
